power: add hardreset action
Devices which do not have a persistent console, cannot pass an
active connection to PDUReboot. As PDUReboot will reset the device,
causing the connection to drop which causes the PDUReboot action to
fail.

Introduce a HardReset action to perform the actual PDU reboot. PDUReboot
now becomes a wrapper around HardReset so that all previous
functionality is unchanged.

Change-Id: I3ff1ff5d2557204146581a12b6fb4dd92f8100a7
Signed-off-by: Tyler Baker <tyler.baker@linaro.org>